Move classes required for communication & some fixes
